refactor: payments feature
