+++ src/features/payments/routers/index.ts
@@ -1,0 +1,4 @@
+import "./payment";
+import "./wallets";
+
+export { paymentsRouter } from "./router";
