--- src/features/wallet/routers/index.ts
@@ -1,4 +1,0 @@
-import "./payment";
-import "./wallets";
-
-export { walletRouter } from "./router";
